requirements.txt
pylms==1.0
python-vlc=3.9.6109
virtualenv==16.4.3
mfrc522==0.6.5
mcp3008==0.1a3
RPi.GPIO==0.6.5
